Extract Parser from Ruliad
